Extract Google Pay state loading
Committed-By-Agent: codex
