+++ paymentsheet/src/main/java/com/stripe/android/checkout/injection/CheckoutControllerComponent.kt
@@ -71,11 +71,13 @@ import com.stripe.android.paymentsheet.repositories.SavedPaymentMethodRepository
 import com.stripe.android.paymentsheet.state.CreateLinkState
 import com.stripe.android.paymentsheet.state.DefaultAnalyticsMetadataFactory
 import com.stripe.android.paymentsheet.state.DefaultCreateLinkState
+import com.stripe.android.paymentsheet.state.DefaultGetGooglePayState
 import com.stripe.android.paymentsheet.state.DefaultLinkAccountStatusProvider
 import com.stripe.android.paymentsheet.state.DefaultPaymentElementLoader
 import com.stripe.android.paymentsheet.state.DefaultPaymentMethodFilter
 import com.stripe.android.paymentsheet.state.DefaultRetrieveCustomerEmail
 import com.stripe.android.paymentsheet.state.DefaultTapToAddAvailabilityFactory
+import com.stripe.android.paymentsheet.state.GetGooglePayState
 import com.stripe.android.paymentsheet.state.LinkAccountStatusProvider
 import com.stripe.android.paymentsheet.state.PaymentElementLoader
 import com.stripe.android.paymentsheet.state.PaymentMethodFilter
@@ -156,6 +158,9 @@ internal interface CheckoutControllerModule {
 
     @Binds
     fun bindsCreateLinkState(impl: DefaultCreateLinkState): CreateLinkState
+
+    @Binds
+    fun bindsGetGooglePayState(impl: DefaultGetGooglePayState): GetGooglePayState
 
     @Binds
     fun bindRetrieveCustomerEmail(impl: DefaultRetrieveCustomerEmail): RetrieveCustomerEmail
