+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/content/EmbeddedPaymentElementViewModelComponent.kt
@@ -36,11 +36,13 @@ import com.stripe.android.paymentsheet.repositories.RealElementsSessionRepositor
 import com.stripe.android.paymentsheet.state.CreateLinkState
 import com.stripe.android.paymentsheet.state.DefaultAnalyticsMetadataFactory
 import com.stripe.android.paymentsheet.state.DefaultCreateLinkState
+import com.stripe.android.paymentsheet.state.DefaultGetGooglePayState
 import com.stripe.android.paymentsheet.state.DefaultLinkAccountStatusProvider
 import com.stripe.android.paymentsheet.state.DefaultPaymentElementLoader
 import com.stripe.android.paymentsheet.state.DefaultPaymentMethodFilter
 import com.stripe.android.paymentsheet.state.DefaultRetrieveCustomerEmail
 import com.stripe.android.paymentsheet.state.DefaultTapToAddAvailabilityFactory
+import com.stripe.android.paymentsheet.state.GetGooglePayState
 import com.stripe.android.paymentsheet.state.LinkAccountStatusProvider
 import com.stripe.android.paymentsheet.state.PaymentElementLoader
 import com.stripe.android.paymentsheet.state.PaymentMethodFilter
@@ -156,6 +158,11 @@ internal interface EmbeddedPaymentElementViewModelModule {
     fun bindsCreateLinkState(
         impl: DefaultCreateLinkState,
     ): CreateLinkState
+
+    @Binds
+    fun bindsGetGooglePayState(
+        impl: DefaultGetGooglePayState,
+    ): GetGooglePayState
 
     @Binds
     fun bindRetrieveCustomerEmail(
