+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/state/PaymentElementLoader.kt
@@ -1,9 +1,6 @@
 package com.stripe.android.paymentsheet.state
 
 import android.os.Parcelable
-import androidx.annotation.VisibleForTesting
-import com.stripe.android.DefaultCardBrandFilter
-import com.stripe.android.DefaultCardFundingFilter
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.SharedPaymentTokenSessionPreview
 import com.stripe.android.common.analytics.experiment.LogFcLiteExperiment
@@ -18,8 +15,6 @@ import com.stripe.android.core.utils.DurationProvider
 import com.stripe.android.core.utils.FeatureFlag
 import com.stripe.android.core.utils.FeatureFlags
 import com.stripe.android.core.utils.UserFacingLogger
-import com.stripe.android.googlepaylauncher.GooglePayEnvironment
-import com.stripe.android.googlepaylauncher.injection.GooglePayRepositoryFactory
 import com.stripe.android.link.LinkController
 import com.stripe.android.lpmfoundations.paymentmethod.AnalyticsMetadata
 import com.stripe.android.lpmfoundations.paymentmethod.CustomerMetadata
@@ -56,7 +51,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import kotlinx.parcelize.Parcelize
 import javax.inject.Inject
@@ -264,12 +258,12 @@ internal interface PaymentElementLoader {
 @SuppressWarnings("LargeClass")
 internal class DefaultPaymentElementLoader @Inject constructor(
     private val prefsRepositoryFactory: PrefsRepository.Factory,
-    private val googlePayRepositoryFactory: GooglePayRepositoryFactory,
     private val logger: Logger,
     private val eventReporter: LoadingEventReporter,
     private val errorReporter: ErrorReporter,
     @IOContext private val workContext: CoroutineContext,
     private val createLinkState: CreateLinkState,
+    private val getGooglePayState: GetGooglePayState,
     private val logLinkHoldbackExperiment: LogLinkHoldbackExperiment,
     private val logFcLiteExperiment: LogFcLiteExperiment,
     private val externalPaymentMethodsRepository: ExternalPaymentMethodsRepository,
@@ -322,30 +316,25 @@ internal class DefaultPaymentElementLoader @Inject constructor(
         eventReporter.onLoadStarted(metadata.initializedViaCompose)
         tapToAddConnectionStarter.start(configuration)
 
-        // Give immediately available results a chance to complete before later load work checks isCompleted.
-        val isGooglePaySupportedOnDevice = async(start = CoroutineStart.UNDISPATCHED) {
-            durationProvider.measureDuration(
-                DurationProvider.Key.PaymentSheetLoadIsGooglePaySupported
-            ) {
-                isGooglePaySupportedOnDevice()
-            }
-        }
-        val isGooglePaySupportedByConfiguration = async {
-            durationProvider.measureDuration(
-                DurationProvider.Key.PaymentSheetLoadIsGooglePayReady
-            ) {
-                configuration.isGooglePayReady()
-            }
-        }
-
         val prefetchedPaymentMethods = prefetchPaymentMethodsForLegacyEphemeralKey(configuration)
 
         val savedPaymentMethodSelection = retrieveSavedPaymentMethodSelection(configuration)
-        val elementsSession = loadSession(
-            initializationMode = initializationMode,
-            configuration = configuration,
-            savedPaymentMethodSelection = savedPaymentMethodSelection,
-        )
+        val elementsSessionResult = async {
+            loadSession(
+                initializationMode = initializationMode,
+                configuration = configuration,
+                savedPaymentMethodSelection = savedPaymentMethodSelection,
+            )
+        }
+        // Start Google Pay checks immediately so they continue in parallel with the session load.
+        val googlePayState = async(start = CoroutineStart.UNDISPATCHED) {
+            getGooglePayState(
+                configuration = configuration,
+                initializationMode = initializationMode,
+                elementsSession = elementsSessionResult,
+            )
+        }
+        val elementsSession = elementsSessionResult.await()
 
         // Preemptively prepare Integrity asynchronously if needed, as warm up can take
         // a few seconds.
@@ -355,17 +344,10 @@ internal class DefaultPaymentElementLoader @Inject constructor(
 
         fetchPaymentMethodMessaging(elementsSession)
 
-        val isGooglePayReady = isGooglePayReady(
-            configuration = configuration,
-            elementsSession = elementsSession,
-            initializationMode = initializationMode,
-            isGooglePaySupportedByConfiguration = isGooglePaySupportedByConfiguration,
-        )
-
         val savedSelection = async {
             retrieveSavedSelection(
                 configuration = configuration,
-                isGooglePayReady = isGooglePayReady,
+                isGooglePayReady = googlePayState.await().isGooglePayReady,
                 elementsSession = elementsSession
             )
         }
@@ -396,9 +378,7 @@ internal class DefaultPaymentElementLoader @Inject constructor(
 
         val paymentMethodMetadata = async {
             val linkStateResult = linkState.await()
-            val isGooglePaySupported = isGooglePaySupportedOnDevice.completeResultOrNull {
-                errorReporter.report(ErrorReporter.ExpectedErrorEvent.GOOGLE_PAY_SKIPPED_DURING_LOAD)
-            } ?: false
+            val googlePayState = googlePayState.await()
 
             durationProvider.measureDuration(DurationProvider.Key.PaymentSheetLoadComputePaymentMethodTypes) {
                 createPaymentMethodMetadata(
@@ -406,8 +386,8 @@ internal class DefaultPaymentElementLoader @Inject constructor(
                     elementsSession = elementsSession,
                     configuration = configuration,
                     linkStateResult = linkStateResult,
-                    isGooglePayReady = isGooglePayReady,
-                    isGooglePaySupported = isGooglePaySupported,
+                    isGooglePayReady = googlePayState.isGooglePayReady,
+                    isGooglePaySupported = googlePayState.isGooglePaySupported,
                     initializationMode = initializationMode,
                     customerMetadata = customerMetadata,
                     clientAttributionMetadata = clientAttributionMetadata,
@@ -440,7 +420,7 @@ internal class DefaultPaymentElementLoader @Inject constructor(
                     savedSelection = savedSelection,
                     metadata = paymentMethodMetadata,
                     customer = customer,
-                    isGooglePayReady = isGooglePayReady,
+                    isGooglePayReady = googlePayState.await().isGooglePayReady,
                     isUsingWalletButtons = configuration.walletButtons?.willDisplayExternally ?: false
                 )
             }
@@ -638,79 +618,6 @@ internal class DefaultPaymentElementLoader @Inject constructor(
                     integrationConfiguration.configuration.paymentMethodLayout
                 }
         }
-    }
-
-    @VisibleForTesting
-    internal suspend fun isGooglePayReady(
-        configuration: CommonConfiguration,
-        elementsSession: ElementsSession,
-        initializationMode: PaymentElementLoader.InitializationMode,
-        isGooglePaySupportedByConfiguration: Deferred<Boolean>,
-    ): Boolean {
-        val shouldDisableForAutomaticTaxBilling =
-            (initializationMode as? PaymentElementLoader.InitializationMode.CheckoutSession)
-            ?.checkoutSessionResponse
-            ?.let { checkoutSessionResponse ->
-                checkoutSessionResponse.automaticTaxEnabled &&
-                    checkoutSessionResponse.taxAddressSource == CheckoutSessionResponse.TaxAddressSource.BILLING &&
-                    configuration.defaultBillingDetails == null
-            } == true
-
-        if (!elementsSession.isGooglePayEnabled) {
-            userFacingLogger.logWarningWithoutPii(
-                "Google Pay is not enabled for this session."
-            )
-        } else if (configuration.googlePay == null) {
-            userFacingLogger.logWarningWithoutPii(
-                "GooglePayConfiguration is not set."
-            )
-        } else if (shouldDisableForAutomaticTaxBilling) {
-            userFacingLogger.logWarningWithoutPii(
-                "Google Pay is disabled because automatic tax is configured to use the billing address and no " +
-                    "default billing address was provided."
-            )
-            return false
-        } else if (!isGooglePaySupportedByConfiguration.await()) {
-            @Suppress("MaxLineLength")
-            userFacingLogger.logWarningWithoutPii(
-                """
-                    Google Pay API check failed.
-                    Possible reasons:
-                    - Google Play service is not available on this device.
-                    - Google account is not signed in on this device.
-                    See https://developers.google.com/android/reference/com/google/android/gms/wallet/PaymentsClient#public-taskboolean-isreadytopay-isreadytopayrequest-request for more details.
-                """.trimIndent()
-            )
-        }
-        return elementsSession.isGooglePayEnabled && isGooglePaySupportedByConfiguration.await()
-    }
-
-    // Default filters are used here because this only determines the ready state,
-    // not what's presented to Google Pay. This check runs async before we fetch the
-    // elements session, so using merchant-defined filters would add latency.
-    private suspend fun isGooglePayReadyForEnvironment(environment: GooglePayEnvironment): Boolean {
-        return googlePayRepositoryFactory(
-            environment = environment,
-            cardFundingFilter = DefaultCardFundingFilter,
-            cardBrandFilter = DefaultCardBrandFilter
-        ).isReady().first()
-    }
-
-    private suspend fun CommonConfiguration.isGooglePayReady(): Boolean {
-        return googlePay?.environment?.let { environment ->
-            isGooglePayReadyForEnvironment(
-                when (environment) {
-                    PaymentSheet.GooglePayConfiguration.Environment.Production ->
-                        GooglePayEnvironment.Production
-                    PaymentSheet.GooglePayConfiguration.Environment.Test ->
-                        GooglePayEnvironment.Test
-                }
-            )
-        } ?: false
-    }
-
-    private suspend fun isGooglePaySupportedOnDevice(): Boolean {
-        return isGooglePayReadyForEnvironment(GooglePayEnvironment.Production)
     }
 
     @Suppress("CyclomaticComplexMethod")
@@ -921,13 +828,4 @@ internal class DefaultPaymentElementLoader @Inject constructor(
 
 private fun PaymentMethod.toPaymentSelection(): PaymentSelection.Saved {
     return PaymentSelection.Saved(this)
-}
-
-private suspend fun <T> Deferred<T>.completeResultOrNull(
-    skippedCallback: () -> Unit,
-): T? = if (isCompleted) {
-    await()
-} else {
-    skippedCallback()
-    null
 }
